users / single user reducer / combined
