+++ client/store/singleUser.js
@@ -6,33 +6,21 @@ import history from '../history'
  */
 const GET_USER = 'GET_USER'
 const REMOVE_USER = 'REMOVE_USER'
-const GET_ALL_USERS = 'GET_ALL_USERS'
 
 /**
  * INITIAL STATE
  */
-const defaultUser = {
-  allUser: []
-}
+const defaultUser = {}
 
 /**
  * ACTION CREATORS
  */
 const getUser = user => ({type: GET_USER, user})
-const removeUser = id => ({type: REMOVE_USER, userId: id})
-const getAllUsers = allUsers => ({type: GET_ALL_USERS, allUsers})
+const removeUser = () => ({type: REMOVE_USER})
 
 /**
  * THUNK CREATORS
  */
-export const fetchAllUsers = () => async dispatch => {
-  try {
-    const res = await axios.get('/api/users')
-    dispatch(getAllUsers(res.data))
-  } catch (err) {
-    console.error(err)
-  }
-}
 
 export const me = () => async dispatch => {
   try {
@@ -72,16 +60,12 @@ export const logout = () => async dispatch => {
 /**
  * REDUCER
  */
-export default function(state = defaultUser, action) {
+export default function userReducer(state = defaultUser, action) {
   switch (action.type) {
     case GET_USER:
-      return {...state, allUser: action.user}
+      return action.user
     case REMOVE_USER:
-      let currUser = [...state.allUser]
-      currUser = state.allUser.filter(user => user.id !== action.userId)
-      return {...state, allUser: currUser}
-    case GET_ALL_USERS:
-      return {...state, allUser: action.allUsers}
+      return defaultUser
     default:
       return state
   }
